+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+### 5 November 2019
+
+* Add a check to the network request retrier to see if the session expired.
+
 ### 4 November 2019
 
 * CocoaPods 1.8.